refactor: renamed to index.tsx
