+++ frontend/components/viewer/encode-input/index.tsx
