+++ frontend/components/viewer/interaction-tables/index.tsx
